+++ src/collective/eeafaceted/dashboard/setuphandlers.py
@@ -28,32 +28,44 @@ def add_demo_data(context):
         type='DashboardCollection',
         title='Every elements',
         query=[{u'i': u'path',
-                u'o': u'plone.app.querystring.operation.string.absolutePath',
+                u'o': u'plone.app.querystring.operation.string.path',
                 u'v': u''}],
         customViewFields=CUSTOM_VIEW_FIELDS,
-        showNumberOfItems=False)
+        showNumberOfItems=False,
+        tal_condition=u'',
+        roles_bypassing_talcondition=[],
+        sort_on=None,
+        sort_reversed=False)
     api.content.create(
         container=folder,
         type='DashboardCollection',
         title='My elements',
         query=[
             {u'i': u'path',
-                u'o': u'plone.app.querystring.operation.string.absolutePath',
+                u'o': u'plone.app.querystring.operation.string.path',
                 u'v': u''},
             {u'i': u'Creator',
              u'o': u'plone.app.querystring.operation.string.currentUser',
              u'v': u''}],
         customViewFields=CUSTOM_VIEW_FIELDS,
-        showNumberOfItems=False)
+        showNumberOfItems=False,
+        tal_condition=u'',
+        roles_bypassing_talcondition=[],
+        sort_on=None,
+        sort_reversed=False)
     api.content.create(
         container=folder,
         type='DashboardCollection',
         title='Elements to review',
         query=[{u'i': u'review_state',
-                u'o': u'plone.app.querystring.operation.selection.any',
-                u'v': [u'pending']}],
+                u'o': u'plone.app.querystring.operation.selection.is',
+                u'v': u'pending'}],
         customViewFields=CUSTOM_VIEW_FIELDS,
-        showNumberOfItems=True)
+        showNumberOfItems=True,
+        tal_condition=u'',
+        roles_bypassing_talcondition=[],
+        sort_on=None,
+        sort_reversed=False)
     api.content.create(
         container=folder,
         type='DashboardCollection',
@@ -62,7 +74,11 @@ def add_demo_data(context):
                 u'o': u'plone.app.querystring.operation.date.beforeToday',
                 u'v': u''}],
         customViewFields=CUSTOM_VIEW_FIELDS,
-        showNumberOfItems=True)
+        showNumberOfItems=True,
+        tal_condition=u'',
+        roles_bypassing_talcondition=[],
+        sort_on=None,
+        sort_reversed=False)
     # enable faceted and configure
     enableFacetedDashboardFor(folder, show_left_column=False)
     _updateDefaultCollectionFor(folder, default_collection.UID())